requirements.txt
apache-airflow-providers-http
apache-airflow-providers-apache-spark
boto3